frontend(JSAP-39):added new feature inside it
