+++ frontend/src/actions/fetchAccountsAction.js
@@ -1,9 +1,5 @@
-import * as SecureStore from 'expo-secure-store';
-
-export default function fetchAccountsAction() {
+export default function fetchAccountsAction(authToken) {
   return async function (dispatch) {
-    const authToken = await SecureStore.getItemAsync('authToken');
-
     const result = await fetch('http://10.22.18.36:3001/deposit', {
       method: 'GET',
       headers: {
