+++ frontend/src/screens/AccountsScreen.js
@@ -1,10 +1,15 @@
 import React from 'react';
 import { Content } from 'native-base';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
 import AccountCard from '../components/AccountCard';
-import navigationPropTypes from '../helpers/navigationPropTypes';
+import fetchAccountsAction from '../actions/fetchAccountsAction';
 
-export default function AccountsScreen({ navigation }) {
-  const accounts = navigation.getParam('accounts');
+function AccountsScreen({ accounts, fetchAccounts, user }) {
+  React.useEffect(
+    () => { fetchAccounts(user.token); }, [],
+  );
+
   return (
     <>
       <Content>
@@ -15,5 +20,18 @@ export default function AccountsScreen({ navigation }) {
 }
 
 AccountsScreen.propTypes = {
-  navigation: navigationPropTypes.isRequired,
+  user: PropTypes.shape({ token: PropTypes.string.isRequired }).isRequired,
+  accounts: PropTypes.arrayOf(PropTypes.object).isRequired,
+  fetchAccounts: PropTypes.func.isRequired,
 };
+
+const mapStateToProps = (state) => ({
+  accounts: state.fetchAccountsReducer,
+  user: state.userReducer.user,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  fetchAccounts: (token) => { dispatch(fetchAccountsAction(token)); },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(AccountsScreen);
